fix: round 3 multi-agent review findings — security, correctness, concurrency
Security / DoS prevention:
- HttpDownloader.sanitizeToken: strip all control chars (0x00-0x1F, 0x7F), not
  just CR/LF, to prevent null-byte HTTP header injection (CWE-113)
- HttpDownloader.isPrivateOrLocalhost: add multicast and RFC 6598 CGNAT
  (100.64.0.0/10) to SSRF block list; these were checked in SSRF_SAFE_DNS
  but not in the pre-connect static validator, creating an inconsistency
- CiHttpClientFactory.SSRF_SAFE_DNS: add CGNAT 100.64.0.0/10 block
- GitHubActionsDownloader + GitLabCiDownloader: limit ZIP iteration to
  MAX_ZIP_ENTRIES=1000 to prevent CPU exhaustion from zip entry bombs
- IndexCollectorServer.readMap: change totalValues from int to long to
  prevent integer overflow bypassing the MAX_TOTAL_VALUES guard (with
  MAX_MAP_ENTRIES=500,000 per entry, ~4,295 entries each with 500,000
  values would overflow int and bypass the 10M limit check)

Correctness:
- ClassNameTrie.getName: add bounds check; throws IllegalStateException
  on out-of-range id rather than AIOOBE on malformed/truncated index file
- MavenNexusDownloader.resolveLatestVersion: replace unbounded body().string()
  with streaming read capped at MAX_METADATA_BYTES=1MB (OOM DoS prevention)
- MavenNexusDownloader.doDownload: fix double-close and partial-file-on-disk
  bugs when size limit exceeded; use Files.newOutputStream + success flag +
  finally-deleteIfExists instead of manual close + TOCTOU delete
- HttpDownloader.download: delete partial output file on size-limit exception
  (previously only checksum failures cleaned up)
- AutoMojo: add missing rejectClassLevelParallelForLearn() call in the
  auto→learn fallback path (was already present in PrepareMojo); without it,
  class-level parallel execution silently corrupts the dependency index
- ReactorReorderer.buildPredecessors: replace getOrDefault(p, Set.of()) with
  computeIfAbsent; the Set.of() fallback would throw UnsupportedOperationException
  if a module was somehow absent from the predecessor map

Concurrency:
- CumulativeTestCountListener: replace plain HashSet/int/boolean with
  ConcurrentHashMap.newKeySet()/AtomicInteger/AtomicBoolean for safe use under
  Maven -T N parallel builds; fix skip logic to use finished set membership
  rather than fragile positional iteration through session.getProjects()
- ClassOrderingEngine: remove racy size()<100 pre-check on LOGGED_STATE_ERRORS
  and LOGGED_INDEX_ERRORS; rely solely on add() for dedup (the pre-check was
  not atomic with add() and could log >100 entries under fork parallelism)

Configuration safety:
- StateConfiguration: reject NaN and Infinity in setFailureDecay,
  setMethodFailureDecay, setDurationAlpha, setMethodDurationAlpha;
  NaN passed IEEE 754 d<0||d>1 checks and would silently corrupt the EMA
  on next save, wiping all accumulated failure history

+++ test-order-core/src/main/java/me/bechberger/testorder/IndexCollectorServer.java
@@ -901,7 +901,7 @@ public final class IndexCollectorServer implements AutoCloseable {
 			throw new IOException("Invalid map entry count: " + entryCount);
 		}
 		Map<String, Set<String>> map = new HashMap<>(entryCount);
-		int totalValues = 0;
+		long totalValues = 0;
 		for (int i = 0; i < entryCount; i++) {
 			String key = readString(in);
 			int valueCount = in.readInt();
